+++ src/main/java/nextstep/ladder2/domain/player/NxPlayer.java
@@ -1,4 +1,4 @@
-package nextstep.ladder2.domain;
+package nextstep.ladder2.domain.player;
 
 import java.util.Objects;
