+++ src/main/java/nextstep/ladder2/domain/prize/NxPrize.java
@@ -1,4 +1,4 @@
-package nextstep.ladder2.domain;
+package nextstep.ladder2.domain.prize;
 
 public class NxPrize {
     private final String prizeName;
