+++ src/main/java/nextstep/ladder2/domain/prize/NxPrizes.java
@@ -1,4 +1,4 @@
-package nextstep.ladder2.domain;
+package nextstep.ladder2.domain.prize;
 
 import java.util.List;
